refactor: move SingleAxisDetector to a separate file
